refactored control package structure and some controllers into InteractableControl
